Move submit function to Form component.
