+++ lib/components/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import firebase, { signIn, signOut, reference } from '../firebase';
+import firebase, { reference } from '../firebase';
 import userContainer from '../containers/userContainer';
 import Form from './Form';
 import SignIn from './SignIn';
@@ -15,19 +15,6 @@ class App extends Component {
 
   componentDidMount() {
     firebase.auth().onAuthStateChanged(user => this.props.getUser(user));
-  };
-
-  submit(e) {
-    e.preventDefault()
-    let { name, school, company, location } = e.target
-    let userInfo = {
-          name: name.value,
-          school: school.value,
-          company: company.value,
-          location: location.value,
-        };
-     reference.push(userInfo);
-     document.getElementById('JobForm').reset();
   };
 
   fetchData() {
@@ -46,7 +33,7 @@ class App extends Component {
       return (
         <div>
         <h1>Welcome {user.displayName}</h1>
-          <Form submit={(e) => this.submit(e)}/>
+          <Form />
           <Map />
         </div>
       )
