+++ lib/components/Map.js
@@ -9,10 +9,10 @@ export default class Map extends React.Component {
       zoom: 4,
       mapTypeId: 'terrain',
     });
+    
     axios.get('https://maps.googleapis.com/maps/api/geocode/json?address=Mountain+View,+CA&key=AIzaSyDgN9mkMr-d_T6w6Meov0J8hB57sapxjCc')
     .then(function (response) {
       let coordinate = response.data.results[0].geometry.location;
-      console.log(coordinate);
       this.marker = new google.maps.Marker({
         position: coordinate,
         map: map
@@ -29,7 +29,12 @@ export default class Map extends React.Component {
 
   render() {
     return (
-      <div id='map' ref="map">map!</div>
+      <div
+        id='map'
+        ref="map"
+      >
+        Map of Bootcamp Grads
+      </div>
     );
   }
 }
